FEATURE: Missing ToDos (#21)
* feature; better password validation

* feat: page size validation

* feat: generation date added to summary title

* fix: account tag validator

* feat: generation data for summary and better data consistency

* fix: removing row version

* refactor: unit of work flow

* feat: decreasing the min length for account tags

* feat: detecting concurrency possibility with deadlock

* feat: total count for queries

+++ MyFinance-Backend/MyFinance.Contracts/Common/Paginated.cs
@@ -1,9 +1,9 @@
 ﻿namespace MyFinance.Contracts.Common;
 
-public class Paginated<T>(IReadOnlyCollection<T> items, int pageNumber, int pageSize, int totalCount)
+public class Paginated<T>(IReadOnlyCollection<T> items, int pageNumber, int pageSize, long totalCount)
 {
     public IReadOnlyCollection<T> Items { get; init; } = items;
     public int PageNumber { get; init; } = pageNumber;
     public int PageSize { get; init; } = pageSize;
-    public int TotalCount { get; init; } = totalCount;
+    public long TotalCount { get; init; } = totalCount;
 }
